requirements_pw2.txt
numpy
matplotlib
torch
pybm3d
scikit-image
pillow